+++ app/src/main/cpp/packetizer/h265_packetizer.h
@@ -18,7 +18,7 @@ public:
     explicit H265Packetizer(uint8_t itl, int32_t ssrc) : interleave_(itl), ssrc_(ssrc) {};
     ~H265Packetizer() = default;
     int32_t packetizeFrame(
-        int32_t seq, uint32_t timestamp,
+        uint16_t seq, uint32_t timestamp,
         const Frame& src, size_t &src_offset, const NalUnit& current_nal,
         uint8_t* dst, size_t dst_size
     ) const;
